[fix] Android build: add gradle installation when not found

+++ scripts/install-android-sdk-tools.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 
-. ./env-global.sh
+# Load global variables
+. "$(dirname $0)/env-global.sh"
 
-test -e "${ANDROID_SDK_ROOT}" || mkdir -p "${ANDROID_SDK_ROOT}"
+# Make sure variables are set
+if [[ "_" == "_${ANDROID_SDK_ROOT}" ]]; then
+  echo "Please set env variable ANDROID_SDK_ROOT"
+  exit 1
+fi
+if [[ "_" == "_${ANDROID_SDK_TOOLS_ROOT}" ]]; then
+  echo "Please set env variable ANDROID_SDK_TOOLS_ROOT"
+  exit 1
+fi
 
 if [[ ! -d "${ANDROID_SDK_TOOLS_ROOT}/bin" ]]; then
-  echo "Installing Android SDK tools..."
+  echo "Installing Android SDK tools... ${ANDROID_SDK_TOOLS_ROOT}"
   test -e "sdk-tools-linux-${ANDROID_SDK_TOOLS_VERSION}.zip" || wget -kL https://dl.google.com/android/repository/sdk-tools-linux-${ANDROID_SDK_TOOLS_VERSION}.zip
-  test -e "${ANDROID_SDK_TOOLS_ROOT}" || unzip -qq sdk-tools-linux-${ANDROID_SDK_TOOLS_VERSION}.zip -d "${ANDROID_SDK_TOOLS}"
+  # Get parent folder
+  ANDROID_SDK_TOOLS_PARENT=$(dirname $ANDROID_SDK_ROOT)
+  test -e "${ANDROID_SDK_TOOLS_PARENT}" || mkdir -p "${ANDROID_SDK_TOOLS_PARENT}"
+  test -e "${ANDROID_SDK_TOOLS_ROOT}" || unzip -qq sdk-tools-linux-${ANDROID_SDK_TOOLS_VERSION}.zip -d "${ANDROID_SDK_TOOLS_PARENT}"
   test -e "${ANDROID_SDK_TOOLS_ROOT}" && rm "sdk-tools-linux-${ANDROID_SDK_TOOLS_VERSION}.zip"
 fi
 
@@ -22,11 +34,11 @@ yes | sdkmanager --licenses
 mkdir -p ~/.android
 touch ~/.android/repositories.cfg
 
-echo y | sdkmanager "platform-tools" --sdk_root=${ANDROID_SDK_ROOT} | tee sdkmanager.log
+echo y | sdkmanager "platform-tools" "--sdk_root=${ANDROID_SDK_ROOT}" | tee sdkmanager.log
+echo y | sdkmanager "extras;android;m2repository" "--sdk_root=${ANDROID_SDK_ROOT}" | tee -a  sdkmanager.log
+echo y | sdkmanager "extras;google;m2repository" "--sdk_root=${ANDROID_SDK_ROOT}" | tee -a sdkmanager.log
 
-echo y | sdkmanager "extras;android;m2repository" --sdk_root=${ANDROID_SDK_ROOT} | tee -a  sdkmanager.log
-echo y | sdkmanager "extras;google;m2repository" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
-
+# Install build tools
 echo y | sdkmanager "build-tools;23.0.2" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
 echo y | sdkmanager "build-tools;23.0.3" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
 echo y | sdkmanager "build-tools;25.0.2" --sdk_root=${ANDROID_SDK_ROOT}  | tee -a sdkmanager.log
@@ -34,6 +46,7 @@ echo y | sdkmanager "build-tools;27.0.3" --sdk_root=${ANDROID_SDK_ROOT}  | tee -
 echo y | sdkmanager "build-tools;28.0.3" --sdk_root=${ANDROID_SDK_ROOT}  | tee -a sdkmanager.log
 echo y | sdkmanager "build-tools;29.0.2" --sdk_root=${ANDROID_SDK_ROOT}  | tee -a sdkmanager.log
 
+# Install platforms
 echo y | sdkmanager "platforms;android-16" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
 echo y | sdkmanager "platforms;android-21" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
 echo y | sdkmanager "platforms;android-23" --sdk_root=${ANDROID_SDK_ROOT} | tee -a sdkmanager.log
